+++ src/BugsnagUnity/Client.cs
@@ -76,7 +76,9 @@ namespace BugsnagUnity
             {
                 _contextSetManually = true;
             }
-            SceneManager.sceneLoaded += SceneLoaded;
+
+            SetupSceneLoadedBreadcrumbTracking();
+
             Application.logMessageReceivedThreaded += MultiThreadedNotify;
             Application.logMessageReceived += Notify;
             User.PropertyChanged += (obj, args) => { NativeClient.SetUser(User); };
@@ -92,6 +94,14 @@ namespace BugsnagUnity
             catch (System.Exception ex)
             {
                 // Async behavior is not available in a test environment
+            }
+        }
+
+        private void SetupSceneLoadedBreadcrumbTracking()
+        {
+            if (Configuration.IsBreadcrumbTypeEnabled(BreadcrumbType.Navigation))
+            {
+                SceneManager.sceneLoaded += SceneLoaded;
             }
         }
 
@@ -156,13 +166,15 @@ namespace BugsnagUnity
                     Notify(new Exception[] { exception }, exception.HandledState, null, logType);
                 }
             }
-            else if (logType.IsGreaterThanOrEqualTo(Configuration.BreadcrumbLogLevel))
+            else if (Configuration.ShouldLeaveLogBreadcrumb(logType))
             {
                 Breadcrumbs.Leave(logType.ToString(), BreadcrumbType.Log, new Dictionary<string, string> {
-          { "message", condition },
-        });
+                    { "message", condition },
+                });
             }
         }
+
+      
 
         public void BeforeNotify(Middleware middleware)
         {