+++ src/BugsnagUnity/Configuration.cs
@@ -1,5 +1,7 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
+using BugsnagUnity.Payload;
 using UnityEngine;
 
 namespace BugsnagUnity
@@ -26,17 +28,31 @@ namespace BugsnagUnity
         public virtual TimeSpan MaximumLogsTimePeriod { get; set; } = TimeSpan.FromSeconds(1);
 
         public virtual Dictionary<LogType, int> MaximumTypePerTimePeriod { get; set; } = new Dictionary<LogType, int>
-    {
-        { LogType.Assert, 5 },
-        { LogType.Error, 5 },
-        { LogType.Exception, 20 },
-        { LogType.Log, 5 },
-        { LogType.Warning, 5 },
-    };
+        {
+            { LogType.Assert, 5 },
+            { LogType.Error, 5 },
+            { LogType.Exception, 20 },
+            { LogType.Log, 5 },
+            { LogType.Warning, 5 },
+        };
 
         public virtual TimeSpan UniqueLogsTimePeriod { get; set; } = TimeSpan.FromSeconds(5);
 
         public virtual LogType BreadcrumbLogLevel { get; set; } = LogType.Log;
+
+        public virtual bool ShouldLeaveLogBreadcrumb(LogType logType)
+        {
+            return IsBreadcrumbTypeEnabled(BreadcrumbType.Log)
+                && logType.IsGreaterThanOrEqualTo(BreadcrumbLogLevel);
+        }
+
+        public BreadcrumbType[] EnabledBreadcrumbTypes { get; set; }
+
+        public virtual bool IsBreadcrumbTypeEnabled(BreadcrumbType breadcrumbType)
+        {
+            return EnabledBreadcrumbTypes == null ||
+               EnabledBreadcrumbTypes.Contains(breadcrumbType);
+        }
 
         public virtual string ApiKey { get; protected set; }