+++ src/BugsnagUnity/Payload/BreadcrumbType.cs
@@ -1,48 +1,51 @@
+using System;
 namespace BugsnagUnity.Payload
 {
     /// <summary>
     /// Represents all of the possible breadcrumb types that the Bugsnag API supports.
     /// </summary>
+    [Serializable]
+    [Flags]
     public enum BreadcrumbType
     {
         /// <summary>
         /// A breadcrumb with navigation information.
         /// </summary>
-        Navigation,
+        Navigation = 0,
 
         /// <summary>
         /// A breadcrumb with request information.
         /// </summary>
-        Request,
+        Request = 1,
 
         /// <summary>
         /// A breadcrumb with process information.
         /// </summary>
-        Process,
+        Process = 2,
 
         /// <summary>
         /// A breadcrumb with log information.
         /// </summary>
-        Log,
+        Log = 3,
 
         /// <summary>
         /// A breadcrumb with user information.
         /// </summary>
-        User,
+        User = 4,
 
         /// <summary>
         /// A breadcrumb with state information.
         /// </summary>
-        State,
+        State = 5,
 
         /// <summary>
         /// A breadcrumb with error information.
         /// </summary>
-        Error,
+        Error = 6,
 
         /// <summary>
         /// A manually logged breadcrumb.
         /// </summary>
-        Manual,
+        Manual = 7,
     }
 }