+++ test/mobile/features/steps/unity_steps.rb
@@ -1,13 +1,13 @@
 When("I wait for the game to start") do
   # Wait for a fixed time period
   # TODO: PLAT-6655 Remove the Unity splash screen so we don't have to wait so long
-  sleep 10
+  sleep 5
 end
 
 When('I relaunch the Unity app') do
   Maze.driver.launch_app
   # Wait for a fixed time period
-  sleep 10
+  sleep 5
 end
 
 When("I tap the {string} button") do |button|
@@ -33,6 +33,10 @@ When("I tap the {string} button") do |button|
     press_at center, 650
   when "Change scene"
     press_at center, 750
+  when "Disable Breadcrumbs"
+    press_at center, 850
+  when "Start SDK"
+    press_at center, 950
   end
 end